+++ src/services/CommerceTrackingService.php
@@ -414,10 +414,9 @@ class CommerceTrackingService extends Component
             return;
         }
 
-        $settings = $plugin->getSettings();
         $result = $plugin->getBurrowApi()->publishEvents(
-            $settings->baseUrl,
-            $settings->apiKey,
+            $plugin->getBurrowBaseUrl(),
+            $plugin->getBurrowApiKey(),
             $runtimeState,
             [$eventEnvelope]
         );
@@ -848,10 +847,9 @@ class CommerceTrackingService extends Component
             return true;
         }
 
-        $settings = $plugin->getSettings();
         $result = $plugin->getBurrowApi()->publishEvents(
-            $settings->baseUrl,
-            $settings->apiKey,
+            $plugin->getBurrowBaseUrl(),
+            $plugin->getBurrowApiKey(),
             $runtimeState,
             [$eventEnvelope]
         );